requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dictdiffer==0.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
psutil==5.9.8
pydantic==2.6.1
pydantic_core==2.16.2
pypika-tortoise==0.1.6
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.28.0
